Chat_bot/requirements.txt
Flask==2.3.2
langchain==0.0.232
python-dotenv==1.0.0